@@ -1,33 +1,32 @@
 {
-  "name": "sttv",
+  "name": "nacnano-resume",
   "version": "0.0.0",
+  "private": true,
+  "type": "module",
   "scripts": {
-    "dev": "NODE_ENV=development vite",
+    "dev": "vite dev --port=5911",
     "build": "vite build",
-    "test": "jest",
-    "test:watch": "jest --watchAll",
-    "serve": "vite preview"
+    "preview": "vite preview --port=4911",
+    "check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
+    "check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
+    "lint": "prettier --check \"src/**/*.{css,scss}\" && eslint . --max-warnings=0",
+    "format": "prettier --write \"src/**/*.{css,scss}\" && eslint . --fix"
   },
   "devDependencies": {
-    "@sveltejs/vite-plugin-svelte": "^1.0.0-next.30",
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/svelte": "^3.0.3",
-    "@types/jest": "^27.0.2",
-    "autoprefixer": "^10.2.5",
-    "babel-jest": "^27.3.1",
-    "cssnano": "^5.0.8",
-    "jest": "^27.3.1",
-    "postcss": "^8.2.7",
-    "postcss-load-config": "^3.0.1",
-    "svelte": "^3.35.0",
-    "svelte-jester": "^2.1.5",
-    "svelte-preprocess": "^4.6.9",
-    "ts-jest": "^27.0.7",
-    "typescript": "^4.2.3",
-    "vite": "^2.1.0"
+    "@sveltejs/adapter-static": "3.0.1",
+    "@sveltejs/kit": "2.0.6",
+    "autoprefixer": "10.4.16",
+    "postcss": "8.4.33",
+    "sass": "1.69.7",
+    "svelte": "4.2.8",
+    "svelte-check": "3.6.2",
+    "svelte-preprocess": "5.1.3",
+    "tailwindcss": "3.4.1",
+    "tslib": "2.6.2",
+    "typescript": "5.3.3",
+    "vite": "5.0.11"
   },
   "dependencies": {
-    "prettier": "^2.8.1",
-    "tailwindcss": "^2.1.2"
+    "@sveltejs/vite-plugin-svelte": "^3.0.1"
   }
 }
